+++ examples/weather_app/src/main.rs
@@ -3,6 +3,7 @@ mod modules;
 mod shared;
 
 use dotenv::dotenv;
+use modules::{get_location, post_location};
 use ngyn::prelude::*;
 use ngyn_shuttle::{ShuttleApplication, ShuttleNgyn};
 
@@ -12,6 +13,9 @@ use crate::middlewares::notfound_middleware::NotFoundMiddleware;
 async fn main() -> ShuttleNgyn {
     dotenv().ok();
     let mut app = ShuttleApplication::default();
+
+    app.get("/<location>/<city>", async_wrap(get_location));
+    app.post("/", async_wrap(post_location));
 
     app.use_middleware(NotFoundMiddleware::new());
     app.use_interpreter(shared::ResponseInterpreter {});
