--- examples/weather_app/src/modules/app_module.rs
@@ -1,6 +1,0 @@
-use ngyn::prelude::*;
-
-use super::{HomeModule, WeatherModule};
-
-#[module(imports = [HomeModule, WeatherModule])]
-pub struct AppModule {}
