--- examples/weather_app/src/modules/home/home_module.rs
@@ -1,6 +1,0 @@
-use ngyn::macros::module;
-
-use super::home_controller::HomeController;
-
-#[module(controllers = [HomeController])]
-pub struct HomeModule {}
