--- examples/weather_app/src/modules/home/mod.rs
@@ -1,2 +1,0 @@
-pub mod home_controller;
-pub mod home_module;
