--- examples/weather_app/src/modules/weather/weather_module.rs
@@ -1,6 +1,0 @@
-use ngyn::macros::*;
-
-use super::weather_controller::WeatherController;
-
-#[module(controllers = [WeatherController])]
-pub struct WeatherModule {}
